requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.8
pydantic>=2.0.0
pydantic-settings>=2.0.0
passlib>=1.7.4
python-dotenv>=1.0.0

python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
SQLAlchemy>=1.4.41
psycopg2-binary
alembic>=1.13.1